refactor: platform を transport/infra に分割
- internal/platform を inbound HTTP 層の internal/transport（csrf/handler/
  httpratelimit/jwt/middleware）と技術基盤層の internal/infra（db/httpclient/
  logging/redis）に再編
- 全 import パスを新パッケージへ追従
- jwt パッケージ名をディレクトリ名に合わせて jwtmw → jwt に統一
  （golang-jwt は gojwt エイリアスで衝突回避）
- .go-arch-lint.yml の platform コンポーネントを transport/infra の2つに分割し
  各 http 層・合成ルートの mayDependOn を更新
