+++ internal/app/di/market.go
@@ -3,7 +3,7 @@ package di
 
 import (
 	"stock_backend/internal/feature/candles/twelvedata"
-	"stock_backend/internal/platform/httpclient"
+	"stock_backend/internal/infra/httpclient"
 )
 
 // NewMarket はHTTPクライアント付きの完全に設定されたTwelveDataMarketを生成します。
