+++ internal/infra/logging/email.go
