+++ internal/infra/logging/email_test.go
